chore: renamed WDModal to WithdrawDepositModal
